refactor(card-browser): extract multiselect menu code
