+++ AnkiDroid/src/main/java/com/ichi2/anki/browser/CardBrowserFragment.kt
@@ -99,6 +99,7 @@ import com.ichi2.anki.libanki.DeckId
 import com.ichi2.anki.libanki.undoAvailable
 import com.ichi2.anki.libanki.undoLabel
 import com.ichi2.anki.model.CardStateFilter
+import com.ichi2.anki.model.CardsOrNotes.CARDS
 import com.ichi2.anki.model.SelectableDeck
 import com.ichi2.anki.model.SortType
 import com.ichi2.anki.observability.ChangeManager
@@ -109,11 +110,13 @@ import com.ichi2.anki.scheduling.ForgetCardsDialog
 import com.ichi2.anki.scheduling.SetDueDateDialog
 import com.ichi2.anki.snackbar.showSnackbar
 import com.ichi2.anki.ui.attachFastScroller
+import com.ichi2.anki.ui.internationalization.toSentenceCase
 import com.ichi2.anki.undoAndShowSnackbar
 import com.ichi2.anki.utils.ext.getCurrentDialogFragment
 import com.ichi2.anki.utils.ext.ifNotZero
 import com.ichi2.anki.utils.ext.launchCollectionInLifecycleScope
 import com.ichi2.anki.utils.ext.setFragmentResultListener
+import com.ichi2.anki.utils.ext.sharedPrefs
 import com.ichi2.anki.utils.ext.showDialogFragment
 import com.ichi2.anki.utils.ext.visibleItemPositions
 import com.ichi2.anki.utils.hideKeyboard
@@ -121,6 +124,7 @@ import com.ichi2.anki.utils.showDialogFragmentImpl
 import com.ichi2.anki.withProgress
 import com.ichi2.ui.CardBrowserSearchView
 import com.ichi2.utils.TagsUtil.getUpdatedTags
+import com.ichi2.utils.increaseHorizontalPaddingOfOverflowMenuIcons
 import com.ichi2.utils.replaceText
 import kotlinx.coroutines.Job
 import kotlinx.coroutines.launch
@@ -491,21 +495,111 @@ class CardBrowserFragment :
         // partial 'multi-select' menu provider
         menuHost.addMenuProvider(
             object : MenuProvider {
+                val vm get() = activityViewModel
+
+                private fun canPerformCardInfo(): Boolean = vm.selectedRowCount() == 1
+
+                private fun canPerformMultiSelectEditNote(): Boolean = vm.selectedRowCount() == 1
+
                 override fun onCreateMenu(
                     menu: Menu,
                     menuInflater: MenuInflater,
                 ) {
-                    // TODO: extract conditionally inflating the menus
+                    if (!activityViewModel.isInMultiSelectMode) return
+                    menuInflater.inflate(R.menu.card_browser_multiselect, menu)
+                    menu.findItem(R.id.action_flag).subMenu?.setupFlags()
+                    requireContext().increaseHorizontalPaddingOfOverflowMenuIcons(menu)
+                }
+
+                override fun onPrepareMenu(menu: Menu) {
+                    if (!vm.isInMultiSelectMode) return
+
+                    menu.findItem(R.id.action_reschedule_cards).title =
+                        TR.actionsSetDueDate().toSentenceCase(R.string.sentence_set_due_date)
+
+                    menu.findItem(R.id.action_grade_now).title =
+                        TR.actionsGradeNow().toSentenceCase(R.string.sentence_grade_now)
+
+                    val isFindReplaceEnabled = sharedPrefs().getBoolean(getString(R.string.pref_browser_find_replace), false)
+                    menu.findItem(R.id.action_find_replace).apply {
+                        isVisible = isFindReplaceEnabled
+                        title = TR.browsingFindAndReplace().toSentenceCase(R.string.sentence_find_and_replace)
+                    }
+
+                    menu.findItem(R.id.action_undo).setupUndo()
+
+                    menu.findItem(R.id.action_flag).isVisible = vm.hasSelectedAnyRows()
+                    menu.findItem(R.id.action_suspend_card).apply {
+                        title = TR.browsingToggleSuspend().toSentenceCase(R.string.sentence_toggle_suspend)
+                        // TODO: I don't think this icon is necessary
+                        setIcon(R.drawable.ic_suspend)
+                        isVisible = vm.hasSelectedAnyRows()
+                    }
+                    menu.findItem(R.id.action_toggle_bury).apply {
+                        title = TR.browsingToggleBury().toSentenceCase(R.string.sentence_toggle_bury)
+                        isVisible = vm.hasSelectedAnyRows()
+                    }
+                    menu.findItem(R.id.action_mark_card).apply {
+                        title = TR.browsingToggleMark()
+                        setIcon(R.drawable.ic_star_border_white)
+                        isVisible = vm.hasSelectedAnyRows()
+                    }
+                    menu.findItem(R.id.action_change_deck).isVisible = vm.hasSelectedAnyRows()
+                    menu.findItem(R.id.action_reposition_cards).isVisible = vm.hasSelectedAnyRows()
+                    menu.findItem(R.id.action_grade_now).isVisible = vm.hasSelectedAnyRows()
+                    menu.findItem(R.id.action_reschedule_cards).isVisible = vm.hasSelectedAnyRows()
+                    menu.findItem(R.id.action_edit_tags).isVisible = vm.hasSelectedAnyRows()
+                    menu.findItem(R.id.action_reset_cards_progress).isVisible = vm.hasSelectedAnyRows()
+
+                    menu.findItem(R.id.action_export_selected).apply {
+                        this.title =
+                            if (vm.cardsOrNotes == CARDS) {
+                                resources.getQuantityString(
+                                    R.plurals.card_browser_export_cards,
+                                    vm.selectedRowCount(),
+                                )
+                            } else {
+                                resources.getQuantityString(
+                                    R.plurals.card_browser_export_notes,
+                                    vm.selectedRowCount(),
+                                )
+                            }
+                        isVisible = vm.hasSelectedAnyRows()
+                    }
+
+                    menu.findItem(R.id.action_edit_note).isVisible = canPerformMultiSelectEditNote()
+                    menu.findItem(R.id.action_view_card_info).isVisible = canPerformCardInfo()
+
+                    val deleteNoteItem =
+                        menu.findItem(R.id.action_delete_card).apply {
+                            isVisible = vm.hasSelectedAnyRows()
+                        }
+
+                    launchCatchingTask {
+                        deleteNoteItem.apply {
+                            this.title =
+                                resources.getQuantityString(
+                                    R.plurals.card_browser_delete_notes,
+                                    vm.selectedNoteCount(),
+                                )
+                        }
+                    }
                 }
 
                 override fun onMenuItemSelected(menuItem: MenuItem): Boolean {
-                    if (!activityViewModel.isInMultiSelectMode) return false
+                    if (!vm.isInMultiSelectMode) return false
 
                     Timber.d("CardBrowserFragment::onMenuItemSelected")
                     prepareForUndoableOperation()
+
+                    Flag.entries.find { it.ordinal == menuItem.itemId }?.let { flag ->
+                        updateFlagForSelectedRows(flag)
+                        return true
+                    }
+
                     when (menuItem.itemId) {
                         android.R.id.home -> {
-                            activityViewModel.endMultiSelectMode(SingleSelectCause.NavigateBack)
+                            vm.endMultiSelectMode(SingleSelectCause.NavigateBack)
                             return true
                         }
                         R.id.action_delete_card -> {
@@ -552,6 +646,33 @@ class CardBrowserFragment :
                         }
                         R.id.action_find_replace -> {
                             showFindAndReplaceDialog()
+                            return true
+                        }
+                        R.id.action_change_note_type -> {
+                            Timber.i("Menu: Change note type")
+                            vm.requestChangeNoteType()
+                            return true
+                        }
+                        R.id.action_undo -> {
+                            Timber.w("CardBrowser:: Undo pressed")
+                            requireCardBrowserActivity().onUndo()
+                            return true
+                        }
+                        R.id.action_preview_many -> {
+                            requireCardBrowserActivity().onPreview()
+                            return true
+                        }
+                        R.id.action_edit_note -> {
+                            requireCardBrowserActivity().openNoteEditorForCurrentlySelectedNote()
+                            return true
+                        }
+                        R.id.action_view_card_info -> {
+                            requireCardBrowserActivity().displayCardInfo()
+                            return true
+                        }
+                        R.id.action_grade_now -> {
+                            Timber.i("CardBrowser:: Grade now button pressed")
+                            requireCardBrowserActivity().openGradeNow()
                             return true
                         }
                     }
